Moving default error handlers to server/bindErrorHandlers
